refactor(components): implement modular component architecture and ES modules
- Break down monolithic page component into reusable components
- Create SearchBar, AdvocateTable, PaginationInfo, and AdvocateResults components
- Add reusable Banner UI component with variant support
- Convert seed files to ES modules for Next.js best practices
- Add utility functions for phone number formatting
- Update package.json to use tsx for better TypeScript + ES module support
- Organize components with proper index exports
- Maintain all existing functionality while improving code structure

This refactoring improves maintainability, reusability, and follows modern
Next.js patterns with proper component separation and ES module usage.
