+++ src/components/advocates/index.ts
@@ -1,0 +1,4 @@
+export { SearchBar } from './SearchBar';
+export { AdvocateTable } from './AdvocateTable';
+export { PaginationInfo } from './PaginationInfo';
+export { AdvocateResults } from './AdvocateResults';
